Changes nick DB to hash
With big DB file (500K rows), evaluation of nicks spiked CPU use of irssi as it needed to traverse the whole file each time.
I have restructured the DB to be ip-unique and load it into irssi as a hash table for faster seach. Not only the amount of rows is reduced by more than 50%, but we also get faster searches, which should solve the CPU issue.